refactor displaytodo view
